+++ static/app/views/explore/hooks/useTraceItemAttributeKeysOptions.tsx
